+++ lib/config.js
@@ -1,14 +1,13 @@
 const path = require('path')
 const qs = require('querystring')
 const Joi = require('joi')
-const ProcessPlugin = require('./plugins/process_plugin')
-const FsPlugin = require('./plugins/fs_plugin')
 const micromatch = require('micromatch')
 const union = require('lodash.union')
 const merge = require('lodash.merge')
-const BrowserSyncPlugin = require('browser-sync-webpack-plugin')
 const {accessSync} = require('fs')
 const hygienist = require('hygienist-middleware')
+const BrowserSyncPlugin = require('browser-sync-webpack-plugin')
+const SpikePlugin = require('./plugin')
 const SpikeUtils = require('spike-util')
 
 /**
@@ -226,16 +225,15 @@ module.exports = class Config {
 
     const util = new SpikeUtils(this)
 
-    this.plugins = [new FsPlugin(util)]
-      .concat(opts.plugins)
-      .concat([
-        new ProcessPlugin(util),
-        new BrowserSyncPlugin(opts.server, { callback: (_, bs) => {
-          if (bs.utils.devIp.length) {
-            this.project.emit('info', `External IP: http://${bs.utils.devIp[0]}:${this.spike.server.port}`)
-          }
-        } })
-      ])
+    this.plugins = [
+      ...opts.plugins,
+      new SpikePlugin(util),
+      new BrowserSyncPlugin(opts.server, { callback: (_, bs) => {
+        if (bs.utils.devIp.length) {
+          this.project.emit('info', `External IP: http://${bs.utils.devIp[0]}:${this.spike.server.port}`)
+        }
+      } })
+    ]
 
     return this
   }
