+++ lib/plugin.js
@@ -1,5 +1,5 @@
 /**
- * @module FSPlugin
+ * @module SpikePlugin
  */
 
 const path = require('path')
@@ -8,37 +8,101 @@ const File = require('filewrap')
 const difference = require('lodash.difference')
 
 /**
- * @class FsWebpackPlugin
- * @classdesc A webpack plugin that collects files to be parsed from the project
- * root and makes them available to other plugins
+ * @class SpikeWebpackPlugin
+ * @classdesc A webpack plugin that reads files from a Spike project, adds them
+ * to webpack's pipeline, lets them compile, then extracts and writes them with
+ * the correct extension.
  */
-module.exports = class FsWebpackPlugin {
+module.exports = class SpikeWebpackPlugin {
 
   /**
    * @constructor
-   * @param {Object} util - a SpikeUtils object
-   * @param {Object} util.conf - configuration values from webpack + spike
-   * @param {Object} util.conf.spike - spike specific config values
+   * @param {Object} opts - options for configuration
+   * @param {String} opts.root - project root
+   * @param {Array} opts.dumpDirs - directories to dump to public
    */
   constructor (util) {
     this.util = util
   }
 
   /**
-   * Main webpack plugin entry point
+   * Generic webpack plugin hook method
+   * see https://webpack.github.io/docs/plugins.html
+   * @param {Object} compiler - webpack compiler instance
    */
   apply (compiler) {
-    this.util.runAll(compiler, this.run.bind(this, compiler))
+    let processedFiles, filesOnly
+
+    // pull files from the project root and categorize them
+    this.util.runAll(compiler, this.sortFiles.bind(this, compiler))
+
+    // inject files into webpack's pipeline
+    compiler.plugin('make', (compilation, done) => {
+      // stores files to be processed by spike
+      processedFiles = compiler.options.spike.files.process
+      filesOnly = processedFiles.map((f) => f.path)
+      // see spike-util for details on this method
+      this.util.addFilesAsWebpackEntries(compilation, filesOnly)
+        .done(() => done(), done)
+    })
+
+    // grab the sources and dependencies and export them into the right files
+    // have webpack export them into their own files
+    compiler.plugin('emit', (compilation, done) => {
+      processedFiles.forEach((f) => {
+        // find the compiled file in webpack's modules
+        const dep = compilation.modules.find((el) => {
+          if (el.userRequest === f.path) { return el }
+        })
+
+        // get the compiled source of the module
+        // this will either come from the source-loader, which exports a clean
+        // buffer as dep._src, or from webpack's raw source value
+        const src = dep._src
+          ? dep._src
+          : JSON.parse(dep._source._value.replace(/^module\.exports = /, ''))
+
+        // calculate the output path
+        let outputPath = this.util.getOutputPath(f.path).relative
+
+        // set the file's extension if necessary
+        if (f.extension) {
+          outputPath = outputPath.replace(/(.*)?(\..+?$)/, `$1.${f.extension}`)
+        }
+
+        // and hand off the file to webpack for emission!
+        compilation.assets[outputPath] = {
+          source: () => { return src },
+          size: () => { return src.length }
+        }
+      })
+
+      done()
+    })
+
+    // remove assets from webpack pipeline since we already wrote them
+    compiler.plugin('compilation', (compilation) => {
+      compilation.plugin('optimize-chunk-assets', (chunks, done) => {
+        this.util.removeAssets(compilation, filesOnly, chunks)
+        done()
+      })
+    })
   }
 
   /**
-   * Function that is executed at the beginning of the compile process
+   * Executed at the beginning of the compile process, this function reads the
+   * spike project's file tree and sorts the files according to how they will
+   * be processed. Called automatically through a webpack hook.
+   * @param {Object} compiler - webpack's compiler
+   * @param {Object} compilation - webpack's compilation
+   * @param {Function} done - callback
    */
-  run (compiler, compilation, done) {
+  sortFiles (compiler, compilation, done) {
     // first we load up and categorize all the files in the project
     compiler.options.spike.files = this.getFilesFromGlob()
 
-    // then we add files to be processed by core loaders as their tests
+    // now we set the core loaders' "test" properties to a very precise list of
+    // files so that they only load the files we want
     compiler.options.module.loaders.map((l) => {
       if (!l._core) { return }
       l.test = pathToRegex(compiler.options.spike.files[l._core])
@@ -57,7 +121,6 @@ module.exports = class FsWebpackPlugin {
 
     // First, we grab all the files in the project, other than the ignored
     // files of course.
-
     const matcher = path.join(util.conf.context, '**/**')
     files.all = glob.sync(matcher, { ignore: util.conf.spike.ignore, dot: true, nodir: true })
 
@@ -67,7 +130,6 @@ module.exports = class FsWebpackPlugin {
 
     // We start with a list of all files, which will be gradually cut down as
     // we find files to be ignored
-
     let allAvailableFiles = files.all
 
     // First step, we find any files that are already being processed by a
@@ -135,14 +197,14 @@ module.exports = class FsWebpackPlugin {
     // Now with any files that are left over, we add them in without an
     // extension property. This means that they will be processed by spike as
     // static files, but their extensions wont be changed.
-
     const noExtension = allAvailableFiles.map((f) => { return { path: f } })
     files.process.push(...noExtension)
 
     // Also add vendor files, as they are processed by spike's pipeline
     files.process.push(...files.vendor.map((f) => { return { path: f } }))
 
-    // finally, add static/vendor to the static category for the loader
+    // finally, add static/vendor to the "static" category so that our static
+    // loader will process these files
     files.static.push(...allAvailableFiles, ...files.vendor)
 
     return files
